Create chicken_nuggest.py
there are only 3 fixed sizes of chicken nuggets. For any given amount of chicken nuggets can the 
order be filled with any combination of the box of 3, box of 6, or box of 20 chicken nuggets. return true if the order can
be filled, return false if the order cannot be filled