Create D3workshop_ex.py
The workshop will be for 15 participants of similar experience and motivation. To that end we are asking you to demonstrate a base level of familiarity with data and programming in the form below. Syntax and language choice are not important, we just want to insure that all participants are at the same level of comfort with certain abstract concepts so everyone can move faster.